feat(*): improve title UX
This commit removes complexity around displaying and parsing jobs. The
logic is simple. The entire text is stored in the JobAd. The title is
the first paragraph (split by <p> or \n).

Then if there is a title then the title is sanitized and truncated.
Otherwise the entire text is sanitized and truncated.

Now the remote flag is set if text includes "REMOTE" (as opposed to just
the title).

Logic is moved from the model to presentation layer in this case.